fix(#5022): remove unwired relay delivery evidence
Drop the process-local evidence layer that had no production writers and restore dead-frontier recovery coverage, while retaining the independent in-flight emission fences and backstop proof hardening.
