+++ ARCHITECTURE.md
@@ -532,7 +532,6 @@ src/
 │   │   │   ├── runtime_resolve.rs
 │   │   │   ├── session_enrichment.rs
 │   │   │   ├── snapshot.rs
-│   │   │   ├── snapshot_relay.rs
 │   │   │   ├── stall_liveness.rs
 │   │   │   ├── stall_verdict.rs
 │   │   │   └── watcher_respawn.rs
@@ -596,7 +595,6 @@ src/
 │   │   │   ├── confirmation.rs
 │   │   │   ├── decision.rs
 │   │   │   ├── delivery.rs
-│   │   │   ├── delivery_evidence_store.rs
 │   │   │   ├── delivery_frontier_probe.rs
 │   │   │   ├── delivery_record.rs
 │   │   │   ├── manual_delivery.rs
