+++ docs/agent-maintenance/discord-outbound-migration.md
@@ -1,5 +1,3 @@
-> Last refreshed: 2026-07-31 (#5022 relay-health evidence authority; outbound callsite coverage unchanged.)
-
 # Discord Outbound Migration — Coverage Map (#1006 v3 / #1280 / #1436 / #1457)
 
 > Last refreshed: 2026-07-24 (#4508 review follow-up — watcher anchored short-replace now uses an edit-only deferred transport boundary in both controller and retained legacy paths. On edit failure, the range owner keeps the same delivery lease and suppresses fallback POST only after a locked, stable pre-edit-path/generation + EOF-bounded durable-frontier recheck; any rotate, marker, metadata, or frontier uncertainty remains fail-open. Already-committed reconciliation reuses delivered-anchor-aware guarded placeholder cleanup and drops local/orphan tracking only after cleanup commits. This changes A4 replacement authority and lifecycle parity, but adds no v3 producer or direct-send callsite; the coverage rows below are unchanged).
@@ -637,5 +635,3 @@ changing runtime behavior.
 > Last refreshed: 2026-07-24 (#4533 — the optional manual-delivery transcript record explicitly opts out of the interactive `/clear` turn-start fence because it is a synthetic post-delivery pair, not an in-flight user turn. No delivery verb, transport, dedup identity, target grammar, or callsite coverage changed.)
 
 > Last refreshed: 2026-07-29 (#4911/#4961 Phase A R9 — `outbound/delivery_record.rs` gains `WatcherDeliveryRecordAuthority` (frontier token + lease-time reset incarnation + source generation + pinned nonzero ledger user_msg_id) threaded into `record_watcher_terminal_delivery` → `shadow_mirror_delivered_frontier_inner`. Watcher and session-sink terminal deliveries now capture that immutable source identity BEFORE transport and re-verify it inside the frontier mutation guard, so a POST that landed after its source generation/incarnation was replaced advances nothing and records no fingerprint/ledger side effect (reported as `LandedStale`); a landed POST whose durable record write fails is reported as `LandedUnrecorded` instead of a silent `Delivered`. No new delivery verb, transport, target grammar, or dedup identity: the production callsite coverage map is unchanged — the same callsites now route their advance/record/fingerprint epilogue through the single guarded funnel instead of independent ambient writes.)
->
-> Last refreshed: 2026-07-31 (#5022 — process-local `RelayDeliveryEvidence` observes exact terminal lease attempts and outcomes for relay-health classification. The evidence store is a read-only health authority and does not add an outbound delivery verb, transport, retry, or direct-send producer; existing outbound callsite coverage remains unchanged.)
