+++ src/services/discord/health.rs
@@ -34,8 +34,6 @@ mod relay_progress;
 mod runtime_resolve;
 mod session_enrichment;
 mod snapshot;
-#[path = "health/snapshot_relay.rs"]
-mod snapshot_relay;
 mod stall_liveness;
 mod stall_verdict;
 mod watcher_respawn;
