+++ src/services/discord/health/liveness_authority.rs
@@ -646,10 +646,8 @@ mod tests {
                 thread_channel_id: None,
                 last_relay_ts_ms: None,
                 last_outbound_activity_ms: None,
-                delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
                 last_capture_offset: None,
                 last_relay_offset: 0,
-                last_relay_offset_recorded: true,
                 unread_bytes: None,
                 desynced: true,
                 stale_thread_proof: false,
