+++ src/services/discord/health/relay_auto_heal.rs
@@ -847,10 +847,8 @@ mod tests {
                 thread_channel_id: None,
                 last_relay_ts_ms: Some(1_700_000_000_000),
                 last_outbound_activity_ms: None,
-                delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
                 last_capture_offset: Some(capture_offset),
                 last_relay_offset,
-                last_relay_offset_recorded: true,
                 unread_bytes: Some(unread_bytes),
                 desynced: true,
                 stale_thread_proof: false,
@@ -1036,7 +1034,7 @@ mod tests {
         shared
             .tmux_relay_coord(channel_id)
             .confirmed_end_offset
-            .store(256, std::sync::atomic::Ordering::Release);
+            .store(256, Ordering::Release);
 
         assert!(!nudge_existing_watcher_for_backlog(
             &shared,
@@ -1080,9 +1078,7 @@ mod tests {
         );
         let snapshot = backlog_snapshot(channel_id, tmux_session, output_path, 100, 301_613);
         let coord = shared.tmux_relay_coord(channel_id);
-        coord
-            .confirmed_end_offset
-            .store(100, std::sync::atomic::Ordering::Release);
+        coord.confirmed_end_offset.store(100, Ordering::Release);
         let high_token = shared.relay_frontier_token(channel_id);
 
         assert!(coord.reset_confirmed_frontier(100, 0));
@@ -1129,10 +1125,7 @@ mod tests {
         shared
             .tmux_relay_coord(channel_id)
             .confirmed_end_offset
-            .store(
-                snapshot.last_relay_offset,
-                std::sync::atomic::Ordering::Release,
-            );
+            .store(snapshot.last_relay_offset, Ordering::Release);
         // Prime the stall observation, then mark a relay emission in-flight
         // (non-zero `relay_slot`) while the committed frontier stays frozen.
         assert!(!nudge_existing_watcher_for_backlog(
@@ -1372,7 +1365,7 @@ mod tests {
         shared
             .tmux_relay_coord(channel_id)
             .confirmed_end_offset
-            .store(128, std::sync::atomic::Ordering::Release);
+            .store(128, Ordering::Release);
         let mut snapshot = backlog_snapshot(channel_id, tmux_session, &output_path, 128, 301_613);
         let started_at = chrono::Local::now().format("%Y-%m-%d %H:%M:%S").to_string();
         snapshot.inflight_started_at = Some(started_at.clone());
@@ -1507,10 +1500,7 @@ mod tests {
         shared
             .tmux_relay_coord(channel_id)
             .confirmed_end_offset
-            .store(
-                snapshot.last_relay_offset,
-                std::sync::atomic::Ordering::Release,
-            );
+            .store(snapshot.last_relay_offset, Ordering::Release);
         let base = 1_800_000_000;
         assert_eq!(
             gated_nudge(
@@ -1631,10 +1621,9 @@ mod tests {
         let tmux_session = "AgentDesk-codex-4181-reset-cap";
         let snapshot = backlog_snapshot(channel_id, tmux_session, "/tmp/4181.jsonl", 128, 256);
         let coord = shared.tmux_relay_coord(channel_id);
-        coord.confirmed_end_offset.store(
-            snapshot.last_relay_offset,
-            std::sync::atomic::Ordering::Release,
-        );
+        coord
+            .confirmed_end_offset
+            .store(snapshot.last_relay_offset, Ordering::Release);
         let base = 1_800_000_000;
         drive_attempt_state_to_cap(&shared, &provider, channel_id, &snapshot, base);
         assert!(
@@ -1669,10 +1658,9 @@ mod tests {
         let snapshot =
             backlog_snapshot(channel_id, tmux_session, "/tmp/4181-shield.jsonl", 128, 256);
         let coord = shared.tmux_relay_coord(channel_id);
-        coord.confirmed_end_offset.store(
-            snapshot.last_relay_offset,
-            std::sync::atomic::Ordering::Release,
-        );
+        coord
+            .confirmed_end_offset
+            .store(snapshot.last_relay_offset, Ordering::Release);
         let key = shared.redrive_key(&provider, channel_id);
         REDRIVE_PLACEHOLDER_SHIELDS.insert(
             key,
@@ -2103,10 +2091,7 @@ mod tests {
         shared
             .tmux_relay_coord(channel_id)
             .confirmed_end_offset
-            .store(
-                snapshot.last_relay_offset,
-                std::sync::atomic::Ordering::Release,
-            );
+            .store(snapshot.last_relay_offset, Ordering::Release);
         snapshot.watcher_owner_channel_id = Some(owner_channel_id.get());
         snapshot.relay_health.watcher_owner_channel_id = Some(owner_channel_id.get());
         let stale_snapshot_owner = ChannelId::new(4_299_007);
@@ -2189,7 +2174,7 @@ mod tests {
         shared
             .tmux_relay_coord(owner_channel_id)
             .confirmed_end_offset
-            .store(64, std::sync::atomic::Ordering::Release);
+            .store(64, Ordering::Release);
         let second_owner = synthetic_rebind_state(
             &provider,
             owner_channel_id,
