+++ src/services/discord/health/stall_liveness.rs
@@ -1124,10 +1124,8 @@ mod tests {
                 thread_channel_id: None,
                 last_relay_ts_ms: Some(1_700_000_000_000),
                 last_outbound_activity_ms: None,
-                delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
                 last_capture_offset: capture_offset,
                 last_relay_offset: 10,
-                last_relay_offset_recorded: true,
                 unread_bytes: capture_offset.map(|offset| offset.saturating_sub(10)),
                 desynced: true,
                 stale_thread_proof: false,
