+++ src/services/discord/health/stall_liveness/redrive_grace.rs
@@ -851,10 +851,8 @@ mod tests {
                 thread_channel_id: None,
                 last_relay_ts_ms: Some(1_700_000_000_000),
                 last_outbound_activity_ms: None,
-                delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
                 last_capture_offset: Some(capture_offset),
                 last_relay_offset: relay_offset,
-                last_relay_offset_recorded: true,
                 unread_bytes: Some(unread),
                 desynced: true,
                 stale_thread_proof: false,
