+++ src/services/discord/health/stall_verdict.rs
@@ -479,10 +479,8 @@ mod tests {
             thread_channel_id: None,
             last_relay_ts_ms: None,
             last_outbound_activity_ms: None,
-            delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
             last_capture_offset: Some(20),
             last_relay_offset: 10,
-            last_relay_offset_recorded: true,
             unread_bytes: Some(10),
             desynced: true,
             stale_thread_proof: false,
@@ -528,7 +526,6 @@ mod tests {
             inflight_state_present: updated_at.is_some(),
             tmux_session_mismatch: false,
             last_relay_offset: 10,
-            last_relay_offset_recorded: true,
             last_relay_ts_ms: 0,
             reconnect_count: 0,
             last_capture_offset: Some(20),
