+++ src/services/discord/outbound/mod.rs
@@ -10,7 +10,6 @@ pub(in crate::services::discord) mod completed_turn_ledger; // #4564
 pub(crate) mod confirmation;
 pub(crate) mod decision;
 pub(crate) mod delivery;
-pub(in crate::services::discord) mod delivery_evidence_store;
 pub(in crate::services::discord) mod delivery_frontier_probe;
 pub(in crate::services::discord) mod delivery_record; // #3089 B0
 pub(crate) mod manual_delivery;
