+++ src/services/discord/relay_recovery_circuit_breaker.rs
@@ -896,10 +896,8 @@ mod tests {
             thread_channel_id: None,
             last_relay_ts_ms: None,
             last_outbound_activity_ms: None,
-            delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
             last_capture_offset: Some(64),
             last_relay_offset: state.last_watcher_relayed_offset.unwrap_or(0),
-            last_relay_offset_recorded: true,
             unread_bytes: Some(64),
             desynced: true,
             stale_thread_proof: false,
