+++ src/services/discord/router/intake_gate/stale_turn.rs
@@ -381,10 +381,8 @@ mod thread_guard_stale_pure_tests {
             thread_channel_id: None,
             last_relay_ts_ms: None,
             last_outbound_activity_ms: None,
-            delivery_evidence: crate::services::discord::outbound::delivery_evidence_store::RelayDeliveryEvidence::NotDelivered,
             last_capture_offset: None,
             last_relay_offset: 0,
-            last_relay_offset_recorded: true,
             unread_bytes: None,
             desynced,
             stale_thread_proof: false,
